// src/api/generated/model/masterOptionResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * 주머니(Jumoney) API 문서
 * 주머니의 백엔드 API 명세서입니다.
 * OpenAPI spec version: 1.0.0
 */
import type { MasterOptionResponseLogicCode } from './masterOptionResponseLogicCode';

export interface MasterOptionResponse {
  optionId?: number;
  logicCode?: MasterOptionResponseLogicCode;
  content?: string;
  description?: string;
  displayOrder?: number;
  requiresSectorSelection?: boolean;
}